Move to TileUtils service
